Move FFI functions to a separate class
