+++ src/Exception.php
@@ -54,8 +54,8 @@ class Exception extends \Exception
     public function __construct($message = "", $code = 0, \Throwable $previous = null)
     {
         if ($message == "") {
-            $message = "libvips error: " . Config::vips()->vips_error_buffer();
-            Config::vips()->vips_error_clear();
+            $message = "libvips error: " . FFI::vips()->vips_error_buffer();
+            FFI::vips()->vips_error_clear();
         }
 
         Utils::errorLog($message);
