+++ src/GValue.php
@@ -46,7 +46,7 @@ class GValue
     public function __construct()
     {
         # allocate a gvalue on the heap, and make it persistent between requests
-        $this->struct = Config::gobject()->new("GValue", true, true);
+        $this->struct = FFI::gobject()->new("GValue", true, true);
         $this->pointer = \FFI::addr($this->struct);
 
         # GValue needs to be inited to all zero
@@ -60,7 +60,7 @@ class GValue
     public static function toEnum(int $gtype, $value): int
     {
         if (is_string($value)) {
-            $enum_value = Config::vips()->
+            $enum_value = FFI::vips()->
                 vips_enum_from_nick("php-vips", $gtype, $value);
             if ($enum_value < 0) {
                 throw new Exception();
@@ -78,7 +78,7 @@ class GValue
      */
     public static function fromEnum(int $gtype, int $value): string
     {
-        $result = Config::vips()->vips_enum_nick($gtype, $value);
+        $result = FFI::vips()->vips_enum_nick($gtype, $value);
         if ($result === null) {
             throw new Exception("value not in enum");
         }
@@ -88,12 +88,12 @@ class GValue
 
     public function __destruct()
     {
-        Config::gobject()->g_value_unset($this->pointer);
+        FFI::gobject()->g_value_unset($this->pointer);
     }
 
     public function setType(int $gtype): void
     {
-        Config::gobject()->g_value_init($this->pointer, $gtype);
+        FFI::gobject()->g_value_init($this->pointer, $gtype);
     }
 
     public function getType(): int
@@ -113,36 +113,36 @@ class GValue
         $gtype = $this->getType();
 
         switch ($gtype) {
-            case Config::gtypes("gboolean"):
-                Config::gobject()->g_value_set_boolean($this->pointer, $value);
+            case FFI::gtypes("gboolean"):
+                FFI::gobject()->g_value_set_boolean($this->pointer, $value);
                 break;
 
-            case Config::gtypes("gint"):
-                Config::gobject()->g_value_set_int($this->pointer, $value);
+            case FFI::gtypes("gint"):
+                FFI::gobject()->g_value_set_int($this->pointer, $value);
                 break;
 
-            case Config::gtypes("gint64"):
-                Config::gobject()->g_value_set_int64($this->pointer, $value);
+            case FFI::gtypes("gint64"):
+                FFI::gobject()->g_value_set_int64($this->pointer, $value);
                 break;
 
-            case Config::gtypes("guint64"):
-                Config::gobject()->g_value_set_uint64($this->pointer, $value);
+            case FFI::gtypes("guint64"):
+                FFI::gobject()->g_value_set_uint64($this->pointer, $value);
                 break;
 
-            case Config::gtypes("gdouble"):
-                Config::gobject()->g_value_set_double($this->pointer, $value);
+            case FFI::gtypes("gdouble"):
+                FFI::gobject()->g_value_set_double($this->pointer, $value);
                 break;
 
-            case Config::gtypes("gchararray"):
-                Config::gobject()->g_value_set_string($this->pointer, $value);
+            case FFI::gtypes("gchararray"):
+                FFI::gobject()->g_value_set_string($this->pointer, $value);
                 break;
 
-            case Config::gtypes("VipsRefString"):
-                Config::vips()->
+            case FFI::gtypes("VipsRefString"):
+                FFI::vips()->
                     vips_value_set_ref_string($this->pointer, $value);
                 break;
 
-            case Config::gtypes("VipsArrayInt"):
+            case FFI::gtypes("VipsArrayInt"):
                 if (!is_array($value)) {
                     $value = [$value];
                 }
@@ -152,11 +152,11 @@ class GValue
                 for ($i = 0; $i < $n; $i++) {
                     $array[$i] = $value[$i];
                 }
-                Config::vips()->
+                FFI::vips()->
                     vips_value_set_array_int($this->pointer, $array, $n);
                 break;
 
-            case Config::gtypes("VipsArrayDouble"):
+            case FFI::gtypes("VipsArrayDouble"):
                 if (!is_array($value)) {
                     $value = [$value];
                 }
@@ -166,17 +166,17 @@ class GValue
                 for ($i = 0; $i < $n; $i++) {
                     $array[$i] = $value[$i];
                 }
-                Config::vips()->
+                FFI::vips()->
                     vips_value_set_array_double($this->pointer, $array, $n);
                 break;
 
-            case Config::gtypes("VipsArrayImage"):
+            case FFI::gtypes("VipsArrayImage"):
                 if (!is_array($value)) {
                     $value = [$value];
                 }
                 $n = count($value);
-                Config::vips()->vips_value_set_array_image($this->pointer, $n);
-                $array = Config::vips()->
+                FFI::vips()->vips_value_set_array_image($this->pointer, $n);
+                $array = FFI::vips()->
                     vips_value_get_array_image($this->pointer, null);
                 for ($i = 0; $i < $n; $i++) {
                     $image = $value[$i];
@@ -185,7 +185,7 @@ class GValue
                 }
                 break;
 
-            case Config::gtypes("VipsBlob"):
+            case FFI::gtypes("VipsBlob"):
                 # we need to set the blob to a copy of the data that vips_lib
                 # can own and free
                 $n = strlen($value);
@@ -194,34 +194,34 @@ class GValue
                 for ($i = 0; $i < $n; $i++) {
                     $memory[$i] = $value[$i];
                 }
-                Config::vips()->
+                FFI::vips()->
                     vips_value_set_blob_free($this->pointer, $memory, $n);
                 break;
 
             default:
-                $fundamental = Config::gobject()->g_type_fundamental($gtype);
+                $fundamental = FFI::gobject()->g_type_fundamental($gtype);
                 switch ($fundamental) {
-                    case Config::gtypes("GObject"):
-                        Config::gobject()->
+                    case FFI::gtypes("GObject"):
+                        FFI::gobject()->
                             g_value_set_object($this->pointer, $value->pointer);
                         break;
 
-                    case Config::gtypes("GEnum"):
-                        Config::gobject()->g_value_set_enum(
+                    case FFI::gtypes("GEnum"):
+                        FFI::gobject()->g_value_set_enum(
                             $this->pointer,
                             self::toEnum($gtype, $value)
                         );
                         break;
 
-                    case Config::gtypes("GFlags"):
+                    case FFI::gtypes("GFlags"):
                         /* Just set as int.
                          */
-                        Config::gobject()->
+                        FFI::gobject()->
                             g_value_set_flags($this->pointer, $value);
                         break;
 
                     default:
-                        $typeName = Config::gobject()->g_type_name($gtype);
+                        $typeName = FFI::gobject()->g_type_name($gtype);
                         throw new \BadMethodCallException(
                             "gtype $typeName ($gtype) not implemented"
                         );
@@ -242,48 +242,48 @@ class GValue
         $result = null;
 
         switch ($gtype) {
-            case Config::gtypes("gboolean"):
-                $result = Config::gobject()->g_value_get_boolean($this->pointer);
+            case FFI::gtypes("gboolean"):
+                $result = FFI::gobject()->g_value_get_boolean($this->pointer);
                 break;
 
-            case Config::gtypes("gint"):
-                $result = Config::gobject()->g_value_get_int($this->pointer);
+            case FFI::gtypes("gint"):
+                $result = FFI::gobject()->g_value_get_int($this->pointer);
                 break;
 
-            case Config::gtypes("gint64"):
-                $result = Config::gobject()->g_value_get_int64($this->pointer);
+            case FFI::gtypes("gint64"):
+                $result = FFI::gobject()->g_value_get_int64($this->pointer);
                 break;
 
-            case Config::gtypes("guint64"):
-                $result = Config::gobject()->g_value_get_uint64($this->pointer);
+            case FFI::gtypes("guint64"):
+                $result = FFI::gobject()->g_value_get_uint64($this->pointer);
                 break;
 
-            case Config::gtypes("gdouble"):
-                $result = Config::gobject()->g_value_get_double($this->pointer);
+            case FFI::gtypes("gdouble"):
+                $result = FFI::gobject()->g_value_get_double($this->pointer);
                 break;
 
-            case Config::gtypes("gchararray"):
-                $result = Config::gobject()->g_value_get_string($this->pointer);
+            case FFI::gtypes("gchararray"):
+                $result = FFI::gobject()->g_value_get_string($this->pointer);
                 break;
 
-            case Config::gtypes("VipsRefString"):
-                $p_size = Config::vips()->new("size_t[1]");
-                $result = Config::vips()->
+            case FFI::gtypes("VipsRefString"):
+                $p_size = FFI::vips()->new("size_t[1]");
+                $result = FFI::vips()->
                     vips_value_get_ref_string($this->pointer, $p_size);
                 # $p_size[0] will be the string length, but assume it's null
                 # terminated
                 break;
 
-            case Config::gtypes("VipsImage"):
-                $pointer = Config::gobject()->g_value_get_object($this->pointer);
+            case FFI::gtypes("VipsImage"):
+                $pointer = FFI::gobject()->g_value_get_object($this->pointer);
                 $result = new Image($pointer);
                 // get_object does not increment the ref count
                 $result->ref();
                 break;
 
-            case Config::gtypes("VipsArrayInt"):
-                $p_len = Config::vips()->new("int[1]");
-                $pointer = Config::vips()->
+            case FFI::gtypes("VipsArrayInt"):
+                $p_len = FFI::vips()->new("int[1]");
+                $pointer = FFI::vips()->
                     vips_value_get_array_int($this->pointer, $p_len);
                 $result = [];
                 for ($i = 0; $i < $p_len[0]; $i++) {
@@ -291,9 +291,9 @@ class GValue
                 }
                 break;
 
-            case Config::gtypes("VipsArrayDouble"):
-                $p_len = Config::vips()->new("int[1]");
-                $pointer = Config::vips()->
+            case FFI::gtypes("VipsArrayDouble"):
+                $p_len = FFI::vips()->new("int[1]");
+                $pointer = FFI::vips()->
                     vips_value_get_array_double($this->pointer, $p_len);
                 $result = [];
                 for ($i = 0; $i < $p_len[0]; $i++) {
@@ -301,9 +301,9 @@ class GValue
                 }
                 break;
 
-            case Config::gtypes("VipsArrayImage"):
-                $p_len = Config::vips()->new("int[1]");
-                $pointer = Config::vips()->
+            case FFI::gtypes("VipsArrayImage"):
+                $p_len = FFI::vips()->new("int[1]");
+                $pointer = FFI::vips()->
                     vips_value_get_array_image($this->pointer, $p_len);
                 $result = [];
                 for ($i = 0; $i < $p_len[0]; $i++) {
@@ -313,31 +313,31 @@ class GValue
                 }
                 break;
 
-            case Config::gtypes("VipsBlob"):
-                $p_len = Config::vips()->new("size_t[1]");
-                $pointer = Config::vips()->
+            case FFI::gtypes("VipsBlob"):
+                $p_len = FFI::vips()->new("size_t[1]");
+                $pointer = FFI::vips()->
                     vips_value_get_blob($this->pointer, $p_len);
                 $result = \FFI::string($pointer, $p_len[0]);
                 break;
 
             default:
-                $fundamental = Config::gobject()->g_type_fundamental($gtype);
+                $fundamental = FFI::gobject()->g_type_fundamental($gtype);
                 switch ($fundamental) {
-                    case Config::gtypes("GEnum"):
-                        $result = Config::gobject()->
+                    case FFI::gtypes("GEnum"):
+                        $result = FFI::gobject()->
                             g_value_get_enum($this->pointer);
                         $result = self::fromEnum($gtype, $result);
                         break;
 
-                    case Config::gtypes("GFlags"):
+                    case FFI::gtypes("GFlags"):
                         /* Just get as int.
                          */
-                        $result = Config::gobject()->
+                        $result = FFI::gobject()->
                             g_value_get_flags($this->pointer);
                         break;
 
                     default:
-                        $typeName = Config::gobject()->g_type_name($gtype);
+                        $typeName = FFI::gobject()->g_type_name($gtype);
                         throw new \BadMethodCallException(
                             "gtype $typeName ($gtype) not implemented"
                         );
