+++ src/Utils.php
@@ -92,23 +92,23 @@ class Utils
      */
     public static function typeFromName(string $name): int
     {
-        return Config::gobject()->g_type_from_name($name);
+        return FFI::gobject()->g_type_from_name($name);
     }
 
     public static function filenameGetFilename(string $name): string
     {
-        $pointer = Config::vips()->vips_filename_get_filename($name);
+        $pointer = FFI::vips()->vips_filename_get_filename($name);
         $filename = \FFI::string($pointer);
-        Config::glib()->g_free($pointer);
+        FFI::glib()->g_free($pointer);
 
         return $filename;
     }
 
     public static function filenameGetOptions(string $name): string
     {
-        $pointer = Config::vips()->vips_filename_get_options($name);
+        $pointer = FFI::vips()->vips_filename_get_options($name);
         $options = \FFI::string($pointer);
-        Config::glib()->g_free($pointer);
+        FFI::glib()->g_free($pointer);
 
         return $options;
     }
